+++ autogen/cluster.tf
@@ -181,6 +181,10 @@ resource "google_container_cluster" "primary" {
       state    = database_encryption.value.state
     }
   }
+
+  workload_identity_config {
+    identity_namespace = var.identity_namespace
+  }
 {% endif %}
 }