+++ modules/beta-public-cluster/variables.tf
@@ -349,3 +349,10 @@ variable "enable_vertical_pod_autoscaling" {
   description = "Vertical Pod Autoscaling automatically adjusts the resources of pods controlled by it"
   default     = false
 }
+
+variable "identity_namespace" {
+  type        = string
+  description = "Workload Identity namespace"
+  default     = ""
+}
+